requirements.txt
python-telegram-bot
openai
schedule
dotenv
